Fix oscillating creep when dropped energy nearby
If the path to the dropped energy moves outside the detection range and
the normal path goes back into range, the creep will oscillate in and
out of range as long as the energy exists. Changed to just ignore the
energy if the path would take us out of range (as opposed to adding
memory of the dropped energy).